subdomain_scanner/requirements.txt
requests
dnspython
PyYAML
aiohttp
aiodns
